+++ src/main/java/kitchenpos/dto/table/OrderTableChangeNumberOfGuestsRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.dto;
+package kitchenpos.dto.table;
 
 import javax.validation.constraints.Min;
